requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
transformers>=4.35.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
pyyaml>=6.0
streamlit>=1.28.0
plotly>=5.17.0
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sec-edgar-downloader>=5.0.3
